1 - Add Migration to convert applicationUser and Genre to ForignKey Property as make Migration 2 - Add Form Design to Create Gig 3 - make Post Action in Controller 4 - make the viewmodel have a fucntion that give full date time to save in db as (Single Responsibility Principle) 5 - Add Resource folder that contain static messages and labels in project

+++ GigHub/Models/Genre.cs
@@ -5,7 +5,7 @@ namespace GigHub.Models
 {
     public class Genre
     {
-        public int ID { get; set; }
+        public byte ID { get; set; }
 
         [Required]
         [StringLength(255)]
